+++ src/bmdscore/trend.h
@@ -4,6 +4,6 @@
 const int WILCOX_MAX = 100;
 // int imax(int i, int j);
 // void w_init(int m, int n);
-int cwilcox(int k, int m, int n);
+double cwilcox(int k, int m, int n);
 
 #endif